move with_diagnostics to ctx
