+++ lib/beaver/mlir/context.ex
@@ -50,4 +50,39 @@ defmodule Beaver.MLIR.Context do
 
   defdelegate destroy(ctx), to: MLIR.CAPI, as: :mlirContextDestroy
   defdelegate register_translations(ctx), to: MLIR.CAPI, as: :mlirRegisterAllLLVMTranslations
+
+  @type result :: any()
+  @type handler_acc :: any()
+  @spec with_diagnostics(
+          MLIR.Context.t(),
+          (-> result),
+          (MLIR.Diagnostic.t(), handler_acc -> handler_acc)
+          | {(-> handler_acc), (MLIR.Diagnostic.t(), handler_acc -> handler_acc)}
+        ) :: {result, handler_acc}
+
+  @doc """
+  Run the given function with a diagnostic handler attached to the MLIR context.
+  """
+  def with_diagnostics(%MLIR.Context{} = ctx, fun, handler)
+      when is_function(fun, 0) and is_function(handler, 2) do
+    with_diagnostics(%MLIR.Context{} = ctx, fun, {fn -> nil end, handler})
+  end
+
+  def with_diagnostics(%MLIR.Context{} = ctx, fun, {init, handler})
+      when is_function(fun, 0) and is_function(init, 0) and is_function(handler, 2) do
+    {:ok, pid} =
+      GenServer.start(
+        Beaver.Capturer,
+        handler
+      )
+
+    handler_id = Beaver.Capturer.attach(ctx, pid)
+
+    try do
+      {fun.(), Beaver.Capturer.collect(pid)}
+    after
+      Beaver.MLIR.Diagnostic.detach(ctx, handler_id)
+      :ok = GenServer.stop(pid)
+    end
+  end
 end
